+++ CHANGELOG.md
@@ -22,6 +22,7 @@ argument or resolve the type ambiguity by casting to the intended type.
 * #4673 Fix now() constification for VIEWs
 * #4681 Fix compression_chunk_size primary key
 * #4685 Improve chunk exclusion for space dimensions
+* #4696 Report warning when enabling compression on hypertable
 
 **Thanks**
 * @maxtwardowski for reporting problems with chunk exclusion and space dimensions